Update the Update structure to return a model and Cmd
Now is possible to return the next command in the update
function of the modules Search and Companies. A lot of changes
in a lot of files was made to do these changes.

This change add the result of the search in the search model.
In the next step, The search result will be processed
in the search module. It this way the result will be available in the
model facilitating the debug.